Tweaked default directory paths to make process
Tweaked default directory paths to make the process of starting this up on a different system more streamlined. Embellished the test data so that the active tracks, cyclone fields, and parameters file are included. The default test data is now assuming no prior tracks in August 2016, so there should be an exact match if the test data inputs are used, prior=0, and the input parameters used match the defaults.