Minor indexing fix
Fixed a bug in np.where() to index call when loading a cyclone field object that already exists. An error was being thrown -- but this would only impact a run where you are running a new coarser temporal resolution with pre-exiting cyclone field objects (e.g., you ran the code at 3-hr previously, and now want to compare to a 6-hr resolution run).